requirements.txt
Faker==19.2.0
Pillow==10.0.0